refactor: Refactored everything in the curriculum, professional-path and certifications, made pages into server components and separated them into different components, hooks, types and constants
