+++ pathexplorer/src/app/user/basic-info/page.tsx
@@ -3,8 +3,8 @@ import { BasicInfoForm } from './components/BasicInfoForm';
 
 export default function BasicInfoPage() {
   return (
-    <div className="flex-1 overflow-y-auto bg-gray-50 pt-8 pb-16">
-      <div className="max-w-5xl mx-auto px-8 mb-16">
+    <div className="flex-1 overflow-y-auto">
+      <div className="p-8 max-w-6xl mx-auto">
         <PageHeader
           title="Personal Information"
           subtitle="Fill in and verify your personal information, you can modify it whenever you want."
