+++ pathexplorer/src/app/user/certifications/components/expiringCertificationsScroll.tsx
@@ -2,11 +2,29 @@ import { useGetExpiringCertifications } from '../hooks/useGetExpiringCertificati
 import { ExpiringCertificationCard } from './expiringCertificationCard';
 import { Skeleton } from '@/components/ui/skeleton';
 import { ChevronLeft, ChevronRight, AlertCircle } from 'lucide-react';
-import { useRef } from 'react';
+import { useRef, useEffect } from 'react';
+import { useUser } from '@/features/context/userContext';
 
-export function ExpiringCertificationsScroll() {
-  const { expiringCertifications, loading, error } = useGetExpiringCertifications();
+interface ExpiringCertificationsScrollProps {
+  isLoading?: boolean;
+  refreshTrigger?: number;
+}
+
+export function ExpiringCertificationsScroll({
+  isLoading = false,
+  refreshTrigger = 0,
+}: ExpiringCertificationsScrollProps) {
+  const { expiringCertifications, loading: apiLoading, error, refetch } = useGetExpiringCertifications(false); // Don't fetch on mount
   const scrollContainerRef = useRef<HTMLDivElement>(null);
+  const { userAuth } = useUser();
+
+  const loading = isLoading || apiLoading;
+
+  useEffect(() => {
+    if (userAuth?.accessToken) {
+      refetch();
+    }
+  }, [refreshTrigger, userAuth?.accessToken, refetch]);
 
   const scroll = (direction: 'left' | 'right') => {
     if (scrollContainerRef.current) {
@@ -19,81 +37,57 @@ export function ExpiringCertificationsScroll() {
     }
   };
 
-  if (loading) {
-    return (
-      <div className="mt-8">
-        <div className="flex items-center mb-4">
-          <h1 className="text-xl font-semibold">Expiring or expired certifications</h1>
-          <div className="ml-auto flex gap-2">
-            <Skeleton className="h-8 w-8 rounded-full" />
-            <Skeleton className="h-8 w-8 rounded-full" />
+  return (
+    <div className="bg-white rounded-lg border shadow-sm">
+      <div className="p-4 border-b border-gray-200 flex items-center justify-between">
+        <h2 className="text-lg font-semibold">Expiring Certifications</h2>
+        {!loading && expiringCertifications.length > 0 && (
+          <div className="flex gap-2">
+            <button
+              onClick={() => scroll('left')}
+              className="p-1 rounded-full bg-white border border-gray-300 hover:bg-gray-50"
+              aria-label="Scroll left"
+            >
+              <ChevronLeft className="h-3 w-3" />
+            </button>
+            <button
+              onClick={() => scroll('right')}
+              className="p-1 rounded-full bg-white border border-gray-300 hover:bg-gray-50"
+              aria-label="Scroll right"
+            >
+              <ChevronRight className="h-3 w-3" />
+            </button>
           </div>
-        </div>
-        <div className="flex gap-4 overflow-x-auto pb-4">
+        )}
+      </div>
+
+      {loading ? (
+        <div className="p-4 flex space-x-8 overflow-x-auto">
           {Array(4).fill(0).map((_, i) => (
-            <div key={`skeleton-${i}`} className="min-w-[280px]">
-              <Skeleton className="h-40 w-full rounded-lg" />
+            <Skeleton key={`skeleton-${i}`} className="h-40 w-80 flex-shrink-0 rounded-lg" />
+          ))}
+        </div>
+      ) : error ? (
+        <div className="p-6 text-center text-red-500">
+          <AlertCircle className="h-5 w-5 mx-auto mb-2" />
+              Failed to load expiring certifications
+        </div>
+      ) : expiringCertifications.length === 0 ? (
+        <div className="p-6 text-center text-gray-500">
+              No expiring certifications found
+        </div>
+      ) : (
+        <div
+          ref={scrollContainerRef}
+          className="p-4 flex space-x-8 overflow-x-auto"
+        >
+          {expiringCertifications.map(certification => (
+            <div key={certification.certification_id} className="flex-shrink-0 w-80">
+              <ExpiringCertificationCard certification={certification} />
             </div>
           ))}
         </div>
-      </div>
-    );
-  }
-
-  if (error) {
-    return (
-      <div className="mt-8 bg-red-50 border border-red-200 rounded-lg p-4">
-        <div className="flex items-center">
-          <AlertCircle className="h-5 w-5 text-red-500 mr-2" />
-          <p className="text-red-600 font-medium">Failed to load expiring certifications</p>
-        </div>
-      </div>
-    );
-  }
-
-  if (expiringCertifications.length === 0) {
-    return null;
-  }
-
-  return (
-    <div className="mt-8">
-      <div className="flex items-center mb-4">
-        <h1 className="text-xl font-semibold">Expiring or expired certifications</h1>
-        <div className="ml-auto flex gap-2">
-          <button
-            onClick={() => scroll('left')}
-            className="p-2 rounded-full bg-white border border-gray-300 hover:bg-gray-50"
-            aria-label="Scroll left"
-          >
-            <ChevronLeft className="h-4 w-4" />
-          </button>
-          <button
-            onClick={() => scroll('right')}
-            className="p-2 rounded-full bg-white border border-gray-300 hover:bg-gray-50"
-            aria-label="Scroll right"
-          >
-            <ChevronRight className="h-4 w-4" />
-          </button>
-        </div>
-      </div>
-
-      <div
-        ref={scrollContainerRef}
-        className="flex gap-4 overflow-x-auto pb-4 scroll-smooth"
-        style={{ msOverflowStyle: 'none', scrollbarWidth: 'none' }}
-      >
-        <style jsx>{`
-          div::-webkit-scrollbar {
-            display: none;
-          }
-        `}</style>
-        {expiringCertifications.map(certification => (
-          <ExpiringCertificationCard
-            key={certification.certification_id}
-            certification={certification}
-          />
-        ))}
-      </div>
+      )}
     </div>
   );
 }
