+++ pathexplorer/src/app/user/certifications/hooks/useRefreshCertificationsStatus.ts
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useState, useCallback } from 'react';
 import { useUser } from '@/features/context/userContext';
 
 export function useRefreshCertificationStatus() {
@@ -6,7 +6,12 @@ export function useRefreshCertificationStatus() {
   const [error, setError] = useState<string | null>(null);
   const { userAuth } = useUser();
 
-  const refreshStatus = async () => {
+  const refreshStatus = useCallback(async () => {
+    if (!userAuth?.accessToken) {
+      console.error('No access token available for refresh-status');
+      throw new Error('Authentication token not available');
+    }
+
     setLoading(true);
     setError(null);
 
@@ -14,7 +19,7 @@ export function useRefreshCertificationStatus() {
       const response = await fetch('/api/certifications/refresh-status', {
         method: 'POST',
         headers: {
-          'Authorization': `Bearer ${userAuth?.accessToken}`,
+          'Authorization': `Bearer ${userAuth.accessToken}`,
         },
       });
 
@@ -29,7 +34,7 @@ export function useRefreshCertificationStatus() {
     } finally {
       setLoading(false);
     }
-  };
+  }, [userAuth?.accessToken]);
 
   return {
     refreshStatus,
