+++ pathexplorer/src/app/user/certifications/types/CertificationTypes.ts
@@ -1,8 +1,20 @@
+
 export interface CertificationItem {
   name: string;
   type: string;
   description: string;
   url: string;
+}
+
+export interface Certification {
+  name: string;
+  type: string;
+  description: string;
+  certification_date: string;
+  expiration_date: string;
+  certification_id: string;
+  status: string;
+  url?: string;
 }
 
 export interface AICertificationRecommendation {
