+++ pathexplorer/src/constants/constants.ts
@@ -52,3 +52,207 @@ export const signUpSteps = [
   { id: 'contact', name: 'Contact & Security', icon: Lock },
   { id: 'professional', name: 'Professional Info', icon: Briefcase },
 ];
+
+export const techSkillDictionary: Record<string, string[]> = {
+  '.NET Development': [
+    'C# 12 language features',
+    'ASP.NET Core',
+    'Entity Framework Core',
+    'Blazor WebAssembly',
+    '.NET MAUI (cross-platform UI)',
+    'Roslyn analyzers & source generators',
+    'Performance profiling (dotTrace, dotMemory)',
+  ],
+  'AI / Machine Learning Engineering': [
+    'Model Lifecycle',
+    'Vector DBs',
+    'Fine-tuning',
+    'MLOps',
+    'Responsible AI',
+  ],
+  'AI-Augmented Development Skills': [
+    'Prompt engineering',
+    'AI code review',
+    'Workflow orchestration',
+  ],
+  'API & Integration Design': [
+    'REST',
+    'GraphQL',
+    'GRPC',
+    'OpenAPI',
+    'Event Streaming Kafka/Pulsar',
+  ],
+  'AR/VR & 3-D Programming': [
+    'Unity',
+    'Unreal',
+    'WebXR',
+    'Shader programming',
+  ],
+  'AWS Cloud Services': [
+    'AWS Lambda',
+    'Amazon EC2',
+    'Amazon S3',
+    'Amazon RDS',
+    'AWS DynamoDB',
+  ],
+  'Blockchain & Smart-Contract Development': [
+    'Solidity',
+    'Hardhat/Foundry',
+    'security audits',
+  ],
+  'Cloud & DevOps': [
+    'AWS/Azure/GCP',
+    'Kubernetes',
+    'Docker',
+    'IaC',
+    'CI/CD',
+  ],
+  'Collaboration & Product Skills': [
+    'Git mastery',
+    'Agile practices',
+    'Code reviews',
+    'Architectural writing',
+    'Communication',
+  ],
+  'Core CS Foundations': [
+    'Algorithms',
+    'Data structures',
+    'Complexity',
+    'Memory',
+  ],
+  'Data Engineering & Analytics': [
+    'ETL pipelines',
+    'Data modeling',
+    'Data Lake vs Warehouse',
+    'Airflow',
+    'Dbt',
+  ],
+  'Databases & Storage': [
+    'PostgreSQL',
+    'MongoDB',
+    'Cassandra',
+    'Redis',
+    'Elasticsearch',
+  ],
+  'Front-End Frameworks & Libraries': [
+    'React',
+    'Vue',
+    'Angular',
+    'Svelte',
+    'Next.js',
+  ],
+  'Languages & Paradigms': [
+    'JavaScript/TypeScript',
+    'Python',
+    'Java',
+    'Go',
+    'Rust',
+    'functional programming',
+    'OOP',
+  ],
+  'Machine Intelligence & Vision': [
+    'Computer Vision',
+    'NLP',
+    'Transformers',
+    'OpenCV',
+    'YOLO',
+  ],
+  'Microsoft Office & Productivity': [
+    'Excel advanced',
+    'PowerPoint macros',
+    'OneNote organization',
+  ],
+  'Mobile Development': [
+    'iOS (SwiftUI, UIKit)',
+    'Android (Kotlin, Jetpack Compose)',
+    'React Native',
+    'Flutter',
+  ],
+  'Multi-language Fluency': [
+    'Python',
+    'JavaScript',
+    'Java',
+    'C++',
+    'SQL',
+    'Ruby',
+    'PHP',
+    'Swift',
+  ],
+  'Network & Security': [
+    'TCP/IP',
+    'SSL/TLS',
+    'firewalls',
+    'VPN',
+    'OWASP top 10',
+  ],
+  'Performance & Observability': [
+    'Profiling',
+    'Tracing',
+    'Metrics',
+    'Logging',
+    'SRE',
+  ],
+  'Engineering Principles': [
+    'SOLID',
+    'DRY',
+    'KISS',
+    'YAGNI',
+    'Design Patterns',
+  ],
+  'Project & Process Management': [
+    'Scrum',
+    'Kanban',
+    'CI/CD pipelines',
+    'Issue Tracking',
+  ],
+  'QA & Testing': [
+    'Unit testing',
+    'Integration tests',
+    'End-to-end tests',
+    'Mocking',
+    'TDD',
+  ],
+  'Research & Writing Skills': [
+    'Technical writing',
+    'Documentation',
+    'Whitepapers',
+    'Presentations',
+  ],
+  'Systems & Architecture': [
+    'Microservices',
+    'Monolith vs micro',
+    'Event-driven',
+    'CQRS',
+    'Serverless',
+  ],
+  'Technical Design & UX': [
+    'Wireframing',
+    'User Flows',
+    'Accessibility',
+    'Figma',
+  ],
+  'Technical Tools': [
+    'VS Code',
+    'JetBrains IDEs',
+    'Docker Desktop',
+    'GitHub',
+    'Postman',
+  ],
+  'Web Foundations': [
+    'HTML5',
+    'CSS3',
+    'Responsive Design',
+    'Accessibility',
+    'Performance Optimization',
+  ],
+};
+
+export const softOptions = [
+  'Communication',
+  'Teamwork',
+  'Problem Solving',
+  'Adaptability',
+  'Time Management',
+  'Leadership',
+  'Creativity',
+];
